test: fold runtime helper suites
